+++ src/containers/DoctorDetails/DoctorDetails.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 
-import Sidebar from '../Sidebar/Sidebar';
+import Sidebar from '../../components/Sidebar/Sidebar';
 import { setAppointment } from '../../Redux/actions/index';
 import styles from './DoctorDetails.module.css';
